Dissolve contributions feature module (#583)
The contributions/ directory was a pass-through over three foreign
collections (requests, phrases, comments) with no domain rules of its
own. Move each hook to the feature that owns its collection, and
relocate UserContributionsTabs (a UI-only zod schema) into the shared
-contributions.tsx component used by both routes.

- useAnyonesPhraseRequests → features/requests/hooks.ts
- useAnyonesPhrases       → features/phrases/hooks.ts
- useAnyonesComments      → features/comments/hooks.ts
- UserContributionsTabs   → routes/_user/learn/-contributions.tsx
