+++ CLAUDE.md
@@ -339,7 +339,6 @@ const mutation = useMutation({
   - `social/` - Friends, chat, public profiles (`schemas.ts`, `collections.ts`, `live.ts`, `hooks.ts`, `public-profile.ts`)
   - `playlists/` - Playlists & phrase links (`schemas.ts`, `collections.ts`, `hooks.ts`)
   - `feed/` - Activity feed (`schemas.ts`, `hooks.ts`)
-  - `contributions/` - User contributions views (`schemas.ts`, `hooks.ts`)
 - `src/lib/` - Cross-cutting utilities
   - `collections/clear-user.ts` - Clears all user collections on logout
   - `supabase-client.ts` - Supabase client setup
@@ -406,19 +405,18 @@ import { cardReviewsCollection } from './collections'
 
 **Feature domains and what they contain:**
 
-| Domain          | Schemas                                 | Collections                                         | Key Hooks                              |
-| --------------- | --------------------------------------- | --------------------------------------------------- | -------------------------------------- |
-| `profile`       | PublicProfile, MyProfile, LanguageKnown | publicProfiles, myProfile                           | useAuth, useProfile                    |
-| `languages`     | Language, LangTag, LangSchema           | languages, langTags                                 | useLanguageMeta, useLanguageTags       |
-| `phrases`       | PhraseFull, Translation, PhraseSearch   | phrases, phrasesFull (live)                         | useLanguagePhrases, usePhrase          |
-| `deck`          | DeckMeta, CardMeta                      | decks, cards                                        | useDeckMeta, useDeckCards, useDeckPids |
-| `review`        | CardReview, DailyReviewState            | cardReviews, reviewDays                             | useReviewsToday, useReviewMutation     |
-| `requests`      | PhraseRequest                           | phraseRequests                                      | useRequest, useRequestCounts           |
-| `comments`      | RequestComment, CommentPhraseLink       | comments, commentPhraseLinks                        | (inline in components)                 |
-| `social`        | FriendSummary, ChatMessage              | friendSummaries, chatMessages, relationsFull (live) | useRelationFriends, useAllChats        |
-| `playlists`     | PhrasePlaylist, PlaylistPhraseLink      | phrasePlaylists, playlistPhraseLinks                | useOnePlaylist, useLangPlaylists       |
-| `feed`          | FeedActivity                            | (uses React Query)                                  | useFeedLang                            |
-| `contributions` | UserContributionsTabs                   | (uses other collections)                            | useAnyonesPhraseRequests               |
+| Domain      | Schemas                                 | Collections                                         | Key Hooks                              |
+| ----------- | --------------------------------------- | --------------------------------------------------- | -------------------------------------- |
+| `profile`   | PublicProfile, MyProfile, LanguageKnown | publicProfiles, myProfile                           | useAuth, useProfile                    |
+| `languages` | Language, LangTag, LangSchema           | languages, langTags                                 | useLanguageMeta, useLanguageTags       |
+| `phrases`   | PhraseFull, Translation, PhraseSearch   | phrases, phrasesFull (live)                         | useLanguagePhrases, usePhrase          |
+| `deck`      | DeckMeta, CardMeta                      | decks, cards                                        | useDeckMeta, useDeckCards, useDeckPids |
+| `review`    | CardReview, DailyReviewState            | cardReviews, reviewDays                             | useReviewsToday, useReviewMutation     |
+| `requests`  | PhraseRequest                           | phraseRequests                                      | useRequest, useRequestCounts           |
+| `comments`  | RequestComment, CommentPhraseLink       | comments, commentPhraseLinks                        | (inline in components)                 |
+| `social`    | FriendSummary, ChatMessage              | friendSummaries, chatMessages, relationsFull (live) | useRelationFriends, useAllChats        |
+| `playlists` | PhrasePlaylist, PlaylistPhraseLink      | phrasePlaylists, playlistPhraseLinks                | useOnePlaylist, useLangPlaylists       |
+| `feed`      | FeedActivity                            | (uses React Query)                                  | useFeedLang                            |
 
 ### Routing Conventions
