+++ src/features/comments/hooks.ts
@@ -3,12 +3,14 @@ import { eq, useLiveQuery } from '@tanstack/react-db'
 import type { UseLiveQueryResult, uuid } from '@/types/main'
 import type { CommentPhraseLinkType, RequestCommentType } from './schemas'
 import type { PhraseFullFullType } from '@/features/phrases/schemas'
+import type { PhraseRequestType } from '@/features/requests/schemas'
 import {
 	commentPhraseLinksCollection,
 	commentUpvotesCollection,
 	commentsCollection,
 } from './collections'
 import { phrasesFull } from '@/features/phrases/live'
+import { phraseRequestsCollection } from '@/features/requests/collections'
 
 /** Look up a single comment by ID. Returns undefined when no id is given. */
 export const useOneComment = (
@@ -53,3 +55,27 @@ export const useHasCommentUpvote = (commentId: uuid): boolean =>
 				.where(({ upvote }) => eq(upvote.comment_id, commentId)),
 		[commentId]
 	).data?.length
+
+export function useAnyonesComments(
+	uid: uuid,
+	lang?: string
+): UseLiveQueryResult<
+	{ comment: RequestCommentType; request: PhraseRequestType }[]
+> {
+	return useLiveQuery(
+		(q) => {
+			let query = q
+				.from({ comment: commentsCollection })
+				.where(({ comment }) => eq(comment.uid, uid))
+				.join(
+					{ request: phraseRequestsCollection },
+					({ comment, request }) => eq(comment.request_id, request.id),
+					'inner'
+				)
+			if (lang) query = query.where(({ request }) => eq(request.lang, lang))
+
+			return query.orderBy(({ comment }) => comment.created_at, 'desc')
+		},
+		[lang, uid]
+	)
+}
