--- src/features/contributions/index.ts
@@ -1,12 +1,0 @@
-// Feature: contributions — User contributions views
-// Public API for the contributions domain
-
-// Schemas & types
-export { UserContributionsTabs } from './schemas'
-
-// Hooks
-export {
-	useAnyonesPhraseRequests,
-	useAnyonesPhrases,
-	useAnyonesComments,
-} from './hooks'
