--- src/features/contributions/schemas.ts
@@ -1,7 +1,0 @@
-import * as z from 'zod'
-
-export const UserContributionsTabs = z.object({
-	contributionsTab: z
-		.enum(['requests', 'phrases', 'playlists', 'answers', 'comments'])
-		.optional(),
-})
