+++ src/features/phrases/index.ts
@@ -32,6 +32,7 @@ export {
 	useLanguagePhrasesSearch,
 	useLanguagePhrase,
 	useAllMyPhrasesLang,
+	useAnyonesPhrases,
 	usePhrasePlaylists,
 	usePhraseComments,
 	usePhraseProvenance,
