+++ src/routes/_user/learn/$lang.contributions.tsx
@@ -2,9 +2,8 @@ import { createFileRoute } from '@tanstack/react-router'
 import languages from '@/lib/languages'
 import { phraseRequestsCollection } from '@/features/requests/collections'
 
-import { UserContributions } from './-contributions'
+import { UserContributions, UserContributionsTabs } from './-contributions'
 import type { CSSProperties } from 'react'
-import { UserContributionsTabs } from '@/features/contributions/schemas'
 
 export const Route = createFileRoute('/_user/learn/$lang/contributions')({
 	validateSearch: UserContributionsTabs,
